DiskIO Success Tests
Due to the nature of these functions, the pass ability depends on
whether or not you actually select a file from something (I had
curlassTest.arr and .cfg on my desktop and used those).
FOUND ERRORS AND ISSUES:
~Does not recognize .cfg files as what it wants, and creates a new
version of the file with an extra .cfg at the end (so I keep ending up
with cutlassTest.cfg.cfg)
~RequestUserFileSource changes the id slightly upon returning, not sure
if this is wanted but it doesn't pass the last test at the moment.
~ Testing this is strange...