Gross fix to undo/redo problem
Apparenlty there's no way to change an undoable edit event's
significance once it's created, so I had to do this:
~MADE ControlledUndoManager, which had a settable boolean
~Upon reviving events, it wraps them in a hidden editable event class,
using the current value of its boolean to set the significance the of
wrapped event.
~Whenever a highlight is called, the undo manager's boolean is changed
to false, making all the highlight-related changes insignificant.
Afterwords it's changed back.
~One problem, however, is that the undo manager has a stack limit. I
overrides this up, but until the highlights are more efficient, we're
effectively keeping a monstrous stack.